Adding a design contribution section
I added a "Design contributions" section to invite contributors. For that purpose, I created new sections named "contributions" and moved the code part to a newly separated area.

The code content was not touched, but I divided the "Keeping in Touch" section into two as the document did not introduce the code part and went directly to join the slack chat and the setup. And finally, I moved the translation section to the end and replaced some emojis with others that I considered more accurate with the headings' ideas.