Selectively revert suspect commits causing API instability (#852)
* Revert "Fix related media endpoint (#808)"

This reverts commit d7962f17e002dd68821efc2fcb86f7edfc2bcec6.

* Selectively revert "Add logging and fix all logging contexts (#790)"

This reverts commit ce335722fe24ab574b463e227b965eab85621d43.

* Revert "Refactor search controller for consistency and clarity (#778)"

Preserves https://github.com/WordPress/openverse-api/pull/730

This reverts commit d26df92f04120e7382d042ae0b439fb6ea9209be.

* Fix divergent Elasticsearch configurations

In https://github.com/WordPress/openverse-api/pull/712 Elasticsearch configuration was moved out of the search controller area into the Django settings module. This unifies access to ES from across the application and is part of the refactoring being done to support the ES versioned index, multi-stage deployment. This commit keeps those changes and applies them to the old search_controller.py module that did not exist at the time the configuration change was applied.

* Lint

* Prevent circular import

We can replace the Any back with the serializer type in the future.

* Add back related media endpoint integration tests

* Lint