Replace always() calls with more appropriate alternatives
always() will (understandably) *always* run, which means that it will also run if the job is cancelled. For a few cases (e.g. cleanup) we still want this, but for most the alternative GitHub recommends (https://docs.github.com/en/actions/learn-github-actions/expressions#always) or a different set of conditions is more desirable.