Implementation Plan: Filtering and designating results with sensitive textual content
This IP is still pending an initial experiment to see whether the reindex to create the filtered index can be done in a timely manner. I suspect it will and I was eager to get this idea written out to confirm that, provided reindexing is viable, the rest of the solution is coherent.