Create lists in reverse order, walking back in time from last event
This was previously iterating over the events as they were defined, meaning the most recent event was the "labeled" event in most cases. This caused problems when trying to test that a ready-then-draft-then-ready PR which straddled the threshhold used the latest ready date, since all the events were in reverse chronological order.