Use local hook for ESLint pre-commit
Includes moving ESLint dependencies to the workspace root and hositing all configurations (except the test/unit one, as it is covered in a separate PR) into the root configuration. This also prepares for ESLint "flat config" changes as it unifies the configuration in a way that more closely mirrors how the flat config works. Hositing ESLint configuration to the root along with its dependencies removes the burden of ESLint dependencies on each individual package (if we add additional packages, but also in automations, which does not define any eslint dependencies despite having a "root" config that extends the parent root configuration [which works separately from how ESLint merges cascading configurations]) and makes it possible to share a single ESLint configuration for all JavaScript in all our projects without ever needing to duplicate dependencies or configuration variables.