Add new tests using storybook and semantic checks
Set `pointer-events-none` on the svg because otherwise it intercepts clicks. We can bypass this issue in Playwright by using `force: true` on the click event when the checkbox is checked... but as the Playwright docs themselves say, this would bypass a meaningful check, and would make our code ignore the potential for the element actually NOT being clickable. The checks exist to make sure that an element clicked by Playwright is actually able to be clicked by a use (i.e., not hidden, not overlapped, etc). The correct thing to do is fix it by setting pointer events to none.