Fix flaky VCollectionHeader snapshot tests (#4559)
* Try to update all snapshots

Why did this not work???

* Force snapshot update using lowered diff tolerance

The text was such a small portion of the screen that it would only fail _sometimes_ on the default maxPixelDiffRatio but only for the smallest screens, where the text takes up comparatively more of the pixels. On larger screens, it seems to never fail. I will create a follow up issue for this.