Fix task dependencies, allow overriding of index suffix
The TaskGroups that created new indices were returning the name of the index from the taskgroup, in order to allow accessing the index names later in other parts of the DAG. Unfortunately this was breaking the expected dependency list in the DAG; only the returned task was being set upstream to downstream tasks, rather than the last task in the TaskGroup.

TL;DR it was possible for, for example, creation of the filtered index to fail, but then table/index promotion would proceed anyway!

This commit fixes that by pulling the tasks to generate the index names out of the TaskGroups so they can be used directly at any point in the DAG.